fix: 將 userController 中業務邏輯調整至 userService 中、移除多餘的 oauthClient 創建、prisma schema 進行 format 處理
